Add files via upload
I added a bubble plot that shows the details of the donations to each candidate. The bubbles are colored by whether the donations were made by an individual or an organization, the size of the bubble is proportional to the amount of donation. Hovering over the bubble shows the name of the Donor and their occupation