Added bubble plot for details of donations
I made a bubble plot for the Shiny app which shows the details of the donations. Each bubble is a donation, color-coded by whether it was by an individual or an organization. The size of the bubble is proportional to the amount of donation. Hovering over the bubble displays the Candidate Name, amount of donation, Donor name, and donor occupation.